Get frontend tests working again
The code sandbox version used CRA as a base, which has a default Jest configuration. Webpacker on the other hand does not handle this for us, so we're left to our own devices. Babel has to be configured to compile ESM into CommonJS modules and CSS modules require an object proxy to work.  moduleNameMapper would also require more extensions added in the future if we ever import images but for this project that likely won't be needed.